refactor(kubernautagent): split config.go into types and validation files
Extracts all Config/RuntimeConfig/LLMRuntimeConfig/InteractiveConfig/etc.
type declarations plus their JWT-provider validation and EffectiveLLM/
EffectivePhaseConfig helpers into config_types.go. Reduces config.go from
837 to 379 lines (Load/LoadLLMRuntime/Validate/DefaultConfig orchestration
only), no behavioral changes.
